requirements.txt
psycopg2-binary~=2.9.6
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login~=0.6.3
werkzeug
Flask~=3.1.1
SQLAlchemy~=2.0.9
bootstrap-flask
python-dotenv
Flask-marshmallow
marshmallow-sqlalchemy
pytest
pytest-cov
beautifulsoup4
factory-boy~=3.3.0
